encode: Ignore .log files in find_audio_files
FileType.AUDIO.parse seems to, at least sometimes, treat log files as audio files.
This is a hacky solution, and it would be better to fix the logic in AUDIO.parse instead.
However, a log file will always be present here, and it will never have audio,
so this method of dealing with it at least should not pose more problems.